js files are now read from a template
