--- tests/deploy.js
@@ -1,67 +1,0 @@
-// geth --networkid 123 --nodiscover --maxpeers 0 --genesis ./genesis_block.json --datadir ./data  js deploy.js 2>>out.log.geth
-
-console.log("unlocking account");
-personal.unlockAccount(
-    web3.eth.accounts[0],
-    "Write here a good, randomly generated, passphrase!"
-);
-
-function checkWork() {
-    if (!eth.mining) {
-       setTimeout(function() {
-            miner.stop(0);
-        }, 3000);
-        miner.start(1);
-    } else {
-        miner.stop(0);
-    }
-   // if (eth.getBlock("pending").transactions.length > 0) {
-   //     if (eth.mining) return;
-   //     console.log("== Pending transactions! Mining...");
-   //     miner.start(1);
-   // } else {
-   //     miner.stop(0);  // This param means nothing
-   //     console.log("== No transactions! Mining stopped.");
-   // }
-}
-
-var _defaultServiceProvider = web3.eth.accounts[0];
-
-var daoContract = web3.eth.contract([{"constant":true,"inputs":[{"name":"","type":"uint256"}],"name":"proposals","outputs":[{"name":"recipient","type":"address"},{"name":"amount","type":"uint256"},{"name":"description","type":"string"},{"name":"votingDeadline","type":"uint256"},{"name":"openToVote","type":"bool"},{"name":"proposalPassed","type":"bool"},{"name":"proposalHash","type":"bytes32"},{"name":"proposalDeposit","type":"uint256"},{"name":"newServiceProvider","type":"bool"},{"name":"creator","type":"address"}],"type":"function"},{"constant":false,"inputs":[{"name":"_spender","type":"address"},{"name":"_value","type":"uint256"}],"name":"approve","outputs":[{"name":"success","type":"bool"}],"type":"function"},{"constant":true,"inputs":[],"name":"rewardAccount","outputs":[{"name":"","type":"address"}],"type":"function"},{"constant":false,"inputs":[{"name":"_tokenHolder","type":"address"}],"name":"buyTokenProxy","outputs":[{"name":"success","type":"bool"}],"type":"function"},{"constant":true,"inputs":[],"name":"daoCreator","outputs":[{"name":"","type":"address"}],"type":"function"},{"constant":true,"inputs":[],"name":"totalSupply","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[{"name":"_proposalID","type":"uint256"}],"name":"numberOfVotes","outputs":[{"name":"_numberOfVotes","type":"uint256"}],"type":"function"},{"constant":false,"inputs":[{"name":"_proposalID","type":"uint256"},{"name":"_newServiceProvider","type":"address"}],"name":"confirmNewServiceProvider","outputs":[{"name":"_success","type":"bool"}],"type":"function"},{"constant":false,"inputs":[{"name":"_proposalID","type":"uint256"},{"name":"_transactionBytecode","type":"bytes"}],"name":"executeProposal","outputs":[{"name":"_success","type":"bool"}],"type":"function"},{"constant":false,"inputs":[{"name":"_from","type":"address"},{"name":"_to","type":"address"},{"name":"_value","type":"uint256"}],"name":"transferFrom","outputs":[{"name":"success","type":"bool"}],"type":"function"},{"constant":true,"inputs":[],"name":"totalRewardToken","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[],"name":"closingTime","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":false,"inputs":[],"name":"refund","outputs":[],"type":"function"},{"constant":false,"inputs":[{"name":"_recipient","type":"address"},{"name":"_amount","type":"uint256"},{"name":"_description","type":"string"},{"name":"_transactionBytecode","type":"bytes"},{"name":"_debatingPeriod","type":"uint256"},{"name":"_newServiceProvider","type":"bool"}],"name":"newProposal","outputs":[{"name":"_proposalID","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[{"name":"_owner","type":"address"}],"name":"balanceOf","outputs":[{"name":"balance","type":"uint256"}],"type":"function"},{"constant":false,"inputs":[{"name":"_recipient","type":"address"}],"name":"addAllowedAddress","outputs":[{"name":"_success","type":"bool"}],"type":"function"},{"constant":true,"inputs":[],"name":"proposalDeposit","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[],"name":"serviceProvider","outputs":[{"name":"","type":"address"}],"type":"function"},{"constant":true,"inputs":[],"name":"numberOfProposals","outputs":[{"name":"_numberOfProposals","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[],"name":"minValue","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[],"name":"rewards","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"}],"name":"payedOut","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":false,"inputs":[{"name":"_to","type":"address"},{"name":"_value","type":"uint256"}],"name":"transfer","outputs":[{"name":"success","type":"bool"}],"type":"function"},{"constant":true,"inputs":[{"name":"","type":"uint256"}],"name":"allowedRecipients","outputs":[{"name":"","type":"address"}],"type":"function"},{"constant":false,"inputs":[{"name":"_proposalID","type":"uint256"},{"name":"_supportsProposal","type":"bool"}],"name":"vote","outputs":[{"name":"_voteID","type":"uint256"}],"type":"function"},{"constant":false,"inputs":[],"name":"getMyReward","outputs":[],"type":"function"},{"constant":true,"inputs":[{"name":"","type":"address"}],"name":"rewardToken","outputs":[{"name":"","type":"uint256"}],"type":"function"},{"constant":true,"inputs":[{"name":"_owner","type":"address"},{"name":"_spender","type":"address"}],"name":"allowance","outputs":[{"name":"remaining","type":"uint256"}],"type":"function"},{"constant":false,"inputs":[],"name":"halfMinQuorum","outputs":[{"name":"_success","type":"bool"}],"type":"function"},{"constant":false,"inputs":[{"name":"_proposalDeposit","type":"uint256"}],"name":"changeProposalDeposit","outputs":[],"type":"function"},{"constant":true,"inputs":[{"name":"_proposalID","type":"uint256"},{"name":"_recipient","type":"address"},{"name":"_amount","type":"uint256"},{"name":"_transactionBytecode","type":"bytes"}],"name":"checkProposalCode","outputs":[{"name":"_codeChecksOut","type":"bool"}],"type":"function"},{"constant":false,"inputs":[],"name":"receiveDAOReward","outputs":[{"name":"","type":"bool"}],"type":"function"},{"constant":true,"inputs":[],"name":"funded","outputs":[{"name":"","type":"bool"}],"type":"function"},{"inputs":[{"name":"_defaultServiceProvider","type":"address"},{"name":"_daoCreator","type":"address"},{"name":"_minValue","type":"uint256"},{"name":"_closingTime","type":"uint256"}],"type":"constructor"},{"anonymous":false,"inputs":[{"indexed":true,"name":"_from","type":"address"},{"indexed":true,"name":"_to","type":"address"},{"indexed":false,"name":"_value","type":"uint256"}],"name":"Transfer","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"_owner","type":"address"},{"indexed":true,"name":"_spender","type":"address"},{"indexed":false,"name":"_value","type":"uint256"}],"name":"Approval","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"value","type":"uint256"}],"name":"Funded","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"to","type":"address"},{"indexed":false,"name":"amount","type":"uint256"}],"name":"SoldToken","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"to","type":"address"},{"indexed":false,"name":"value","type":"uint256"}],"name":"Refund","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"proposalID","type":"uint256"},{"indexed":false,"name":"recipient","type":"address"},{"indexed":false,"name":"amount","type":"uint256"},{"indexed":false,"name":"description","type":"string"}],"name":"ProposalAdded","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"name":"proposalID","type":"uint256"},{"indexed":false,"name":"position","type":"bool"},{"indexed":true,"name":"voter","type":"address"}],"name":"Voted","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"proposalID","type":"uint256"},{"indexed":false,"name":"result","type":"bool"},{"indexed":false,"name":"quorum","type":"uint256"}],"name":"ProposalTallied","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"_newServiceProvider","type":"address"}],"name":"NewServiceProvider","type":"event"},{"anonymous":false,"inputs":[{"indexed":true,"name":"_recipient","type":"address"}],"name":"AllowedRecipientAdded","type":"event"}]
-);
-
-console.log("Creating DAOCreator Contract");
-var creatorContract = web3.eth.contract([{"constant":false,"inputs":[{"name":"_defaultServiceProvider","type":"address"},{"name":"_minValue","type":"uint256"},{"name":"_closingTime","type":"uint256"}],"name":"createDAO","outputs":[{"name":"_newDAO","type":"address"}],"type":"function"}]
-);
-var _daoCreatorContract = creatorContract.new({from: web3.eth.accounts[0], data: '606060405261229e806100126000396000f3606060405260e060020a60003504630b6d8d528114601a575b005b607a6004356024356044356000833084846060612207806100978339018085600160a060020a0316815260200184600160a060020a03168152602001838152602001828152602001945050505050604051809103906000f0949350505050565b60408051600160a060020a03929092168252519081900360200190f36060604052604051608080612207833960e060408190529151905160a05160c051600c819055600d82905560048054600160a060020a03199081168617909155600b8054909116841790556801158e460913d00000600a559293919290913090610152806100cb8339600160a060020a03909216908201526040519081900361010001906000f060088054600160a060020a0319169190911790819055426002556005600155600160a060020a03166000141560b9576002565b50505050611fea8061021d6000396000f3606060405260405160208061015283395060806040525160008054600160a060020a031916821790555061011b806100376000396000f3606060405236156100355760e060020a60003504630221038a81146100425780638da5cb5b14610074578063d2cc718f14610086575b61008f6001805434019055565b6100916004356024356000805433600160a060020a03908116911614158061006a5750600034115b156100ad57610002565b6100a3600054600160a060020a031681565b61009160015481565b005b60408051918252519081900360200190f35b6060908152602090f35b600160a060020a0383168183606082818181858883f19350505050156101155760408051600160a060020a03851681526020810184905281517f9735b0cb909f3d21d5c16bbcccd272d85fa11446f6d679f6ecb170d2dabfecfc929181900390910190a15060015b92915050566060604052361561017f5760e060020a6000350463013cf08b811461019c578063095ea7b31461030e5780630e7082031461038357806313d4bc2414610395578063149acf9a1461048a57806318160ddd1461049c5780631fcc81be146104a557806322b22bac146104ec578063237e94921461050b57806323b872dd1461057257806334145808146105a65780634b6753bc146105af578063590e1ae3146105b8578063612e45a3146105c957806370a082311461066857806381788e2b1461068e5780638b15a605146106a45780638d69e95e146106ad5780638d7af473146106bf578063963e63c7146106ca5780639ec5a894146106d3578063a79f7ac9146106dc578063a9059cbb146106f4578063b124e88e1461072e578063c9d27afe14610774578063cc9ae3f61461078a578063cdef91d01461079f578063dd62ed3e146107b7578063dd7abd41146107eb578063e33734fd14610818578063eceb29451461082c578063f15f46771461088b578063f3a504f21461089e575b6108aa600c546000906234bc00014211156108bc576108c561088f565b6108cb60043560008054829081101561000257508052600c02600080516020611fca8339815191528101547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5668201547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5648301547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5678401547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5688501547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5698601547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56a8701547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56e880154600160a060020a039788169895977f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56596909601969560ff86811696610100900481169594931691168a565b6108aa60043560243533600160a060020a03908116600081815260106020908152604080832094871680845294825280832086905580518681529051929493927f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925929181900390910190a35060015b92915050565b6109b4600854600160a060020a031681565b6108aa6004355b60006000600c60005054421080156103b45750600034115b15611b805750600160a060020a0382166000818152600f60209081526040918290208054349081019091556011805482019055825181815292519093927f8338662358e4188b6716fe38d845443fe23e7dc9dcb10134fd37212b629d142992908290030190a2600d60005054601160005054101580156104375750600e5460ff16155b1561047f57600e805460ff1916600117905560115460408051918252517fc4c14883ae9fd8e26d5d59e3485ed29fd126d781d7e498a4ca5c54c8268e49369181900360200190a15b600191505b50919050565b6109b4600b54600160a060020a031681565b6108aa60115481565b6108aa6004356000600060005082815481101561000257505080527f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56c600c82020154610689565b6108aa6004356024356000600060006000600034111561150157610002565b60408051602060248035600481810135601f81018590048502860185019096528585526108aa958135959194604494929390920191819084018382808284375094965050505050505060006000600060006000600060006000600034111561115e57610002565b6108aa600435602435604435600e5460009060ff1680156105945750600c5442115b8015611b855750611b85848484610721565b6108aa60075481565b6108aa600c5481565b6109d06000341115611c7157610002565b604080516020604435600481810135601f81018490048402850184019095528484526108aa948135946024803595939460649492939101918190840183828082843750506040805160209735808a0135601f81018a90048a0283018a01909352828252969897608497919650602491909101945090925082915084018382808284375094965050933593505060a435915050600060006109d23361066f565b6108aa6004355b600160a060020a0381166000908152600f60205260409020545b919050565b6108aa60043560006000341115611bf657610002565b6108aa600a5481565b6109b4600454600160a060020a031681565b6108aa6000546108c8565b6108aa600d5481565b6108aa60035481565b6108aa60043560096020526000908152604090205481565b6108aa600435602435600e5460009060ff1680156107135750600c5442115b8015611b575750611b573384845b60006000611d688561066f565b6109b460043560058054829081101561000257506000527f036b6384b5eca791c62761152d0c79bb0604c104a5fb6f4eb0703f3154bb3db00154600160a060020a031681565b6108aa60043560243560006000610fbf3361066f565b6109d06000600060003411156119fb57610002565b6108aa60043560066020526000908152604090205481565b6108aa600435602435600160a060020a0382811660009081526010602090815260408083209385168352929052205461037d565b6108aa60006301dfe20042036002600050541015611c6957504260029081556001805490910281556108c8565b6109d06004356000341115611bba57610002565b604080516020606435600481810135601f81018490048402850184019095528484526108aa94813594602480359560443595608494920191908190840183828082843750949650505050505050600060006000341115610f3057610002565b6108aa5b600380543401905560016108c8565b6108aa600e5460ff1681565b60408051918252519081900360200190f35b6108c53361039c565b90505b90565b604051808b600160a060020a031681526020018a81526020018060200189815260200188815260200187815260200186815260200185815260200184815260200183600160a060020a0316815260200182810382528a81815460018160011615610100020316600290048152602001915080546001816001161561010002031660029004801561099c5780601f106109715761010080835404028352916020019161099c565b820191906000526020600020905b81548152906001019060200180831161097f57829003601f168201915b50509b50505050505050505050505060405180910390f35b60408051600160a060020a039092168252519081900360200190f35b005b600014156109df57610002565b828015610a2957508660001415806109f957508451600014155b80610a115750600454600160a060020a038981169116145b80610a1c5750600034115b80610a29575062093a8084105b15610a3357610002565b82158015610aa75750610a99886004546000908190600160a060020a0380851691161480610a6f5750600854600160a060020a03908116908416145b80610a8b575030600160a060020a031683600160a060020a0316145b15611dcf5760019150610484565b1580610aa757506212750084105b15610ab157610002565b600e5460ff161580610ac45750600c5442105b80610ad95750600a5434108015610ad9575082155b15610ae357610002565b600854600160a060020a038981169116148015610b01575060035487115b15610b0b57610002565b6000805460018101808355909190828015829011610b4257600c0281600c028360005260206000209182019101610b429190610bf9565b505060008054929450918491508110156100025750808052600c8302600080516020611fca8339815191520190508054600160a060020a031916891781556001818101899055875160028084018054600082815260209081902096975091959481161561010002600019011691909104601f908101829004840193918b0190839010610d3a57805160ff19168380011785555b50610d6a929150610cda565b5050600b81018054600160a060020a03191690556001015b80821115610cee578054600160a060020a03191681556000600182810182905560028084018054848255909281161561010002600019011604601f819010610cc057505b506000600383018190556004808401805461ffff19169055600584018290556006840182905560078401805460ff191690556008840180548382559083526020909220610cf2929091028101905b80821115610cee5760008082556001820181815560028301919091556003919091018054600160a060020a0319169055610c8b565b601f016020900490600052602060002090810190610c3d91905b80821115610cee5760008155600101610cda565b5090565b5060098201805460008083559182526020909120610be1918101905b80821115610cee57805474ffffffffffffffffffffffffffffffffffffffffff19168155600101610d0e565b82800160010185558215610bd5579182015b82811115610bd5578251826000505591602001919060010190610d4c565b50508787866040518084600160a060020a0316606060020a0281526014018381526020018280519060200190808383829060006004602084601f0104600f02600301f150905001935050505060405180910390208160050160005081905550834201816003016000508190555060018160040160006101000a81548160ff02191690830217905550828160070160006101000a81548160ff021916908302179055508215610e505760088101805460018101808355909190828015829011610e4b57600402816004028360005260206000209182019101610e4b9190610c8b565b505050505b600b81018054600160a060020a0319163317905534600682015560408051838152600160a060020a038a1660208281019190915291810189905260806060820181815289519183019190915288517f646fec02522b41e7125cfc859a64fd4f4cefd5dc3b6237ca0abe251ded1fa8819386938d938d938d9360a0840191858201918190849082908590600090600490601f850104600f02600301f150905090810190601f168015610f155780820380516001836020036101000a031916815260200191505b509550505050505060405180910390a1509695505050505050565b6000805487908110156100025750808052600c8702600080516020611fca83398151915201905090508484846040518084600160a060020a0316606060020a0281526014018381526020018280519060200190808383829060006004602084601f0104600f02600301f15090500193505050506040518091039020816005016000505414915050949350505050565b60001415610fcc57610002565b6000341115610fda57610002565b600080548590811015610002575033600160a060020a03168152600c85027f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56d8101602052604090912054600080516020611fca83398151915291909101915060ff168061104b575060038101544210155b1561105557610002565b60098101805460018101808355909190828015829011611088578183600052602060002091820191016110889190610d0e565b50506040805180820190915285815233602082015260098401805493955090929091508490811015610002579060005260206000209001600050805482516020938401516101000274ffffffffffffffffffffffffffffffffffffffff001960ff19938416909217919091161790915533600160a060020a03166000818152600a850184526040908190208054939093166001179092558151878152928301869052815190927f86abfce99b7dd908bec0169288797f85049ec73cbe046ed9de818fab3a497ae092908290030190a25092915050565b600080548b9081101561000257508052600c8a027f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e566810154600080516020611fca8339815191529190910197504210806111bd5750600487015460ff16155b806111cc5750600787015460ff165b8061124d57508660000160009054906101000a9004600160a060020a031687600101600050548a6040518084600160a060020a0316606060020a0281526014018381526020018280519060200190808383829060006004602084601f0104600f02600301f15090500193505050506040518091039020876005016000505414155b1561125757610002565b6000955060009450600093505b60098701548410156112a657600987018054859081101561000257600091825260209091200180549093506112c8906101009004600160a060020a031661066f565b506001860154858501906112ef905b6001546011546003830491900401610689565b835490925060ff16156112de57948101946112e3565b938101935b60019390930192611264565b81101580156112fd57508486115b1561133857604051600b8801546006890154600160a060020a0391909116916000919082818181858883f19350505050151561134657610002565b600187015461145f906112b5565b604051600188015488548b51600160a060020a0391909116928c918190602084810191908190849082908590600090600490601f850104600f02600301f150905090810190601f1680156113ae5780820380516001836020036101000a031916815260200191505b5091505060006040518083038185876185025a03f19250505015156113d257610002565b60048701805461ff00191661010017905542600255600854875460019950600160a060020a03908116911614156114155786880154600354101561144b57610002565b60018701805430600160a060020a031660009081526006602052604090208054919091019055546007805490910190555b6114ae565b600187015460038054919091039055611446565b811015801561146e5750858510155b156114ae57604051600b8801546006890154600160a060020a0391909116916000919082818181858883f1935050505015156114a957610002565b426002555b60048701805460ff19169055604080518981526020810183905281518c927fdfc78bdca8e3e0b18c16c5c99323c6cb9eb5e00afde190b4e7273f5158702b07928290030190a25050505050505092915050565b61150a3361066f565b6000141561151757610002565b60008054879081101561000257508052600c86027f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e566810154600080516020611fca83398151915290910193504210806115785750600383015462360d800142115b8061159157508254600160a060020a0390811690861614155b806115a15750600783015460ff16155b156115ab57610002565b600883018054600090811015610002579081526020812060030154600160a060020a03161415611712576117fc85604051600090600160a060020a038316907feaa602bff84cc95bf1b06084485d789e35140c4931eeca8a3b6eb0ed87d74917908390a260408051600b547f0b6d8d52000000000000000000000000000000000000000000000000000000008252600160a060020a0380861660048401526024830185905262375f00420160448401529251921691630b6d8d5291606480820192602092909190829003018187876161da5a03f1156100025750506040515191506106899050565b600883018054600685015430600160a060020a0316310391906000908110156100025781815260208082209390935530600160a060020a03168152600690925260408220548154909290811015610002579081526020812090506002015560115460088401805460009081101561000257908152602081209050600101555b6008830180546000908110156100025781548282526020822060010154929190811015610002579081526020812090505433600160a060020a03166000908152600f602052604081205460088701805493909102939093049450908110156100025790815260208120905060030154604080517f13d4bc2400000000000000000000000000000000000000000000000000000000815233600160a060020a039081166004830152915192909116916313d4bc249185916024828101926020929190829003018185886185025a03f11561000257505060405151600014159150611872905057610002565b60088401805460009081101561000257818152602081206003018054600160a060020a03191690931790925580549091908110156100025790815260208120905060030154600160a060020a03166000141561185757610002565b600683015430600160a060020a031631101561169357610002565b6008830180546000908110156100025781548282526020822060010154929190811015610002579081526020812090506002015433600160a060020a03166000908152600f60205260408120546008870180549390910293909304935083926006928290811015610002575260208120815060030154600160a060020a0390811682526020808301939093526040918201600090812080549095019094553016835260069091529020548190101561192957610002565b30600160a060020a0390811660009081526006602090815260408083208054869003905533909316808352600f82528383205484519081529351929390927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef929181900390910190a3600f600050600033600160a060020a031681526020019081526020016000206000505460116000828282505403925050819055506000600f600050600033600160a060020a03168152602001908152602001600020600050819055506001935050505092915050565b33600160a060020a03818116600090815260066020526040808220546011543090941683529120549092611a2e9061066f565b02040191506009600050600033600160a060020a0316815260200190815260200160002060005054600760005054600860009054906101000a9004600160a060020a0316600160a060020a031663d2cc718f6040518160e060020a0281526004018090506020604051808303816000876161da5a03f1156100025750505060405180519060200150840204039050600860009054906101000a9004600160a060020a0316600160a060020a0316630221038a33836040518360e060020a0281526004018083600160a060020a03168152602001828152602001925050506020604051808303816000876161da5a03f1156100025750506040515115159050611b3557610002565b33600160a060020a031660009081526009602052604090208054820190555050565b8015611b735750611b73838360006000341115611e4157610002565b15611b805750600161037d565b610002565b8015611ba25750611ba284848460006000341115611ee757610002565b15611b8057506001611bb3565b5060005b9392505050565b30600160a060020a031633600160a060020a0316141580611be75750600a30600160a060020a0316310481115b15611bf157610002565b600a55565b600454600160a060020a039081163390911614611c1257610002565b60058054600181018083558281838015829011611c4257818360005260206000209182019101611c429190610cda565b5050506000928352506020909120018054600160a060020a03191683179055506001610689565b5060006108c8565b600c5442118015611c855750600e5460ff16155b8015611cb6575033600160a060020a03166000818152600f602052604080822054905182818181858883f193505050505b15611d665733600160a060020a03166000818152600f602090815260409182902054825190815291517fbb28353e4598c3b9199101a66e0989549b659a59a54d2c27fbb183f1932c8e6d9281900390910190a2600f600050600033600160a060020a031681526020019081526020016000206000505460116000828282505403925050819055506000600f600050600033600160a060020a03168152602001908152602001600020600050819055505b565b600160a060020a038616600090815260096020526040902054808502919091049150811115611d9657610002565b600160a060020a038086166000908152600960205260408082208054859003905591861681522080548201905560019150509392505050565b5060005b600554811015611e3057600580548290811015610002576000919091527f036b6384b5eca791c62761152d0c79bb0604c104a5fb6f4eb0703f3154bb3db00154600160a060020a039081169084161415611e395760019150610484565b60009150610484565b600101611dd3565b33600160a060020a03166000908152600f6020526040902054829010801590611e6a5750600082115b15611edf5733600160a060020a039081166000818152600f6020908152604080832080548890039055938716808352918490208054870190558351868152935191937fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef929081900390910190a350600161037d565b50600061037d565b600160a060020a0384166000908152600f6020526040902054829010801590611f305750601060209081526040600081812033600160a060020a03168252909252902054829010155b8015611f3c5750600082115b15611baf57600160a060020a038084166000818152600f6020908152604080832080548801905584891680845281842080548990039055601083528184203390961684529482529182902080548790039055815186815291519293927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef9281900390910190a3506001611bb356290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e563',
-    gas: 3000000
-       }, function(e, contract){
-           if (e) {
-               console.log(e+" at DAOCreator creation!");
-           }
-           if (typeof contract.address != 'undefined') {
-               console.log('DAOCreator mined! address: ' + contract.address + ' transactionHash: ' + contract.transactionHash);
-               checkWork();
-               var dao = daoContract.new(
-                   _defaultServiceProvider,
-                   contract.address,
-                   50,
-                   1456964290,
-            {
-                from: web3.eth.accounts[0],
-                data: '6060604052604051608080612207833960e060408190529151905160a05160c051600c819055600d82905560048054600160a060020a03199081168617909155600b8054909116841790556801158e460913d00000600a559293919290913090610152806100cb8339600160a060020a03909216908201526040519081900361010001906000f060088054600160a060020a0319169190911790819055426002556005600155600160a060020a03166000141560b9576002565b50505050611fea8061021d6000396000f3606060405260405160208061015283395060806040525160008054600160a060020a031916821790555061011b806100376000396000f3606060405236156100355760e060020a60003504630221038a81146100425780638da5cb5b14610074578063d2cc718f14610086575b61008f6001805434019055565b6100916004356024356000805433600160a060020a03908116911614158061006a5750600034115b156100ad57610002565b6100a3600054600160a060020a031681565b61009160015481565b005b60408051918252519081900360200190f35b6060908152602090f35b600160a060020a0383168183606082818181858883f19350505050156101155760408051600160a060020a03851681526020810184905281517f9735b0cb909f3d21d5c16bbcccd272d85fa11446f6d679f6ecb170d2dabfecfc929181900390910190a15060015b92915050566060604052361561017f5760e060020a6000350463013cf08b811461019c578063095ea7b31461030e5780630e7082031461038357806313d4bc2414610395578063149acf9a1461048a57806318160ddd1461049c5780631fcc81be146104a557806322b22bac146104ec578063237e94921461050b57806323b872dd1461057257806334145808146105a65780634b6753bc146105af578063590e1ae3146105b8578063612e45a3146105c957806370a082311461066857806381788e2b1461068e5780638b15a605146106a45780638d69e95e146106ad5780638d7af473146106bf578063963e63c7146106ca5780639ec5a894146106d3578063a79f7ac9146106dc578063a9059cbb146106f4578063b124e88e1461072e578063c9d27afe14610774578063cc9ae3f61461078a578063cdef91d01461079f578063dd62ed3e146107b7578063dd7abd41146107eb578063e33734fd14610818578063eceb29451461082c578063f15f46771461088b578063f3a504f21461089e575b6108aa600c546000906234bc00014211156108bc576108c561088f565b6108cb60043560008054829081101561000257508052600c02600080516020611fca8339815191528101547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5668201547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5648301547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5678401547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5688501547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e5698601547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56a8701547f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56e880154600160a060020a039788169895977f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56596909601969560ff86811696610100900481169594931691168a565b6108aa60043560243533600160a060020a03908116600081815260106020908152604080832094871680845294825280832086905580518681529051929493927f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925929181900390910190a35060015b92915050565b6109b4600854600160a060020a031681565b6108aa6004355b60006000600c60005054421080156103b45750600034115b15611b805750600160a060020a0382166000818152600f60209081526040918290208054349081019091556011805482019055825181815292519093927f8338662358e4188b6716fe38d845443fe23e7dc9dcb10134fd37212b629d142992908290030190a2600d60005054601160005054101580156104375750600e5460ff16155b1561047f57600e805460ff1916600117905560115460408051918252517fc4c14883ae9fd8e26d5d59e3485ed29fd126d781d7e498a4ca5c54c8268e49369181900360200190a15b600191505b50919050565b6109b4600b54600160a060020a031681565b6108aa60115481565b6108aa6004356000600060005082815481101561000257505080527f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56c600c82020154610689565b6108aa6004356024356000600060006000600034111561150157610002565b60408051602060248035600481810135601f81018590048502860185019096528585526108aa958135959194604494929390920191819084018382808284375094965050505050505060006000600060006000600060006000600034111561115e57610002565b6108aa600435602435604435600e5460009060ff1680156105945750600c5442115b8015611b855750611b85848484610721565b6108aa60075481565b6108aa600c5481565b6109d06000341115611c7157610002565b604080516020604435600481810135601f81018490048402850184019095528484526108aa948135946024803595939460649492939101918190840183828082843750506040805160209735808a0135601f81018a90048a0283018a01909352828252969897608497919650602491909101945090925082915084018382808284375094965050933593505060a435915050600060006109d23361066f565b6108aa6004355b600160a060020a0381166000908152600f60205260409020545b919050565b6108aa60043560006000341115611bf657610002565b6108aa600a5481565b6109b4600454600160a060020a031681565b6108aa6000546108c8565b6108aa600d5481565b6108aa60035481565b6108aa60043560096020526000908152604090205481565b6108aa600435602435600e5460009060ff1680156107135750600c5442115b8015611b575750611b573384845b60006000611d688561066f565b6109b460043560058054829081101561000257506000527f036b6384b5eca791c62761152d0c79bb0604c104a5fb6f4eb0703f3154bb3db00154600160a060020a031681565b6108aa60043560243560006000610fbf3361066f565b6109d06000600060003411156119fb57610002565b6108aa60043560066020526000908152604090205481565b6108aa600435602435600160a060020a0382811660009081526010602090815260408083209385168352929052205461037d565b6108aa60006301dfe20042036002600050541015611c6957504260029081556001805490910281556108c8565b6109d06004356000341115611bba57610002565b604080516020606435600481810135601f81018490048402850184019095528484526108aa94813594602480359560443595608494920191908190840183828082843750949650505050505050600060006000341115610f3057610002565b6108aa5b600380543401905560016108c8565b6108aa600e5460ff1681565b60408051918252519081900360200190f35b6108c53361039c565b90505b90565b604051808b600160a060020a031681526020018a81526020018060200189815260200188815260200187815260200186815260200185815260200184815260200183600160a060020a0316815260200182810382528a81815460018160011615610100020316600290048152602001915080546001816001161561010002031660029004801561099c5780601f106109715761010080835404028352916020019161099c565b820191906000526020600020905b81548152906001019060200180831161097f57829003601f168201915b50509b50505050505050505050505060405180910390f35b60408051600160a060020a039092168252519081900360200190f35b005b600014156109df57610002565b828015610a2957508660001415806109f957508451600014155b80610a115750600454600160a060020a038981169116145b80610a1c5750600034115b80610a29575062093a8084105b15610a3357610002565b82158015610aa75750610a99886004546000908190600160a060020a0380851691161480610a6f5750600854600160a060020a03908116908416145b80610a8b575030600160a060020a031683600160a060020a0316145b15611dcf5760019150610484565b1580610aa757506212750084105b15610ab157610002565b600e5460ff161580610ac45750600c5442105b80610ad95750600a5434108015610ad9575082155b15610ae357610002565b600854600160a060020a038981169116148015610b01575060035487115b15610b0b57610002565b6000805460018101808355909190828015829011610b4257600c0281600c028360005260206000209182019101610b429190610bf9565b505060008054929450918491508110156100025750808052600c8302600080516020611fca8339815191520190508054600160a060020a031916891781556001818101899055875160028084018054600082815260209081902096975091959481161561010002600019011691909104601f908101829004840193918b0190839010610d3a57805160ff19168380011785555b50610d6a929150610cda565b5050600b81018054600160a060020a03191690556001015b80821115610cee578054600160a060020a03191681556000600182810182905560028084018054848255909281161561010002600019011604601f819010610cc057505b506000600383018190556004808401805461ffff19169055600584018290556006840182905560078401805460ff191690556008840180548382559083526020909220610cf2929091028101905b80821115610cee5760008082556001820181815560028301919091556003919091018054600160a060020a0319169055610c8b565b601f016020900490600052602060002090810190610c3d91905b80821115610cee5760008155600101610cda565b5090565b5060098201805460008083559182526020909120610be1918101905b80821115610cee57805474ffffffffffffffffffffffffffffffffffffffffff19168155600101610d0e565b82800160010185558215610bd5579182015b82811115610bd5578251826000505591602001919060010190610d4c565b50508787866040518084600160a060020a0316606060020a0281526014018381526020018280519060200190808383829060006004602084601f0104600f02600301f150905001935050505060405180910390208160050160005081905550834201816003016000508190555060018160040160006101000a81548160ff02191690830217905550828160070160006101000a81548160ff021916908302179055508215610e505760088101805460018101808355909190828015829011610e4b57600402816004028360005260206000209182019101610e4b9190610c8b565b505050505b600b81018054600160a060020a0319163317905534600682015560408051838152600160a060020a038a1660208281019190915291810189905260806060820181815289519183019190915288517f646fec02522b41e7125cfc859a64fd4f4cefd5dc3b6237ca0abe251ded1fa8819386938d938d938d9360a0840191858201918190849082908590600090600490601f850104600f02600301f150905090810190601f168015610f155780820380516001836020036101000a031916815260200191505b509550505050505060405180910390a1509695505050505050565b6000805487908110156100025750808052600c8702600080516020611fca83398151915201905090508484846040518084600160a060020a0316606060020a0281526014018381526020018280519060200190808383829060006004602084601f0104600f02600301f15090500193505050506040518091039020816005016000505414915050949350505050565b60001415610fcc57610002565b6000341115610fda57610002565b600080548590811015610002575033600160a060020a03168152600c85027f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e56d8101602052604090912054600080516020611fca83398151915291909101915060ff168061104b575060038101544210155b1561105557610002565b60098101805460018101808355909190828015829011611088578183600052602060002091820191016110889190610d0e565b50506040805180820190915285815233602082015260098401805493955090929091508490811015610002579060005260206000209001600050805482516020938401516101000274ffffffffffffffffffffffffffffffffffffffff001960ff19938416909217919091161790915533600160a060020a03166000818152600a850184526040908190208054939093166001179092558151878152928301869052815190927f86abfce99b7dd908bec0169288797f85049ec73cbe046ed9de818fab3a497ae092908290030190a25092915050565b600080548b9081101561000257508052600c8a027f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e566810154600080516020611fca8339815191529190910197504210806111bd5750600487015460ff16155b806111cc5750600787015460ff165b8061124d57508660000160009054906101000a9004600160a060020a031687600101600050548a6040518084600160a060020a0316606060020a0281526014018381526020018280519060200190808383829060006004602084601f0104600f02600301f15090500193505050506040518091039020876005016000505414155b1561125757610002565b6000955060009450600093505b60098701548410156112a657600987018054859081101561000257600091825260209091200180549093506112c8906101009004600160a060020a031661066f565b506001860154858501906112ef905b6001546011546003830491900401610689565b835490925060ff16156112de57948101946112e3565b938101935b60019390930192611264565b81101580156112fd57508486115b1561133857604051600b8801546006890154600160a060020a0391909116916000919082818181858883f19350505050151561134657610002565b600187015461145f906112b5565b604051600188015488548b51600160a060020a0391909116928c918190602084810191908190849082908590600090600490601f850104600f02600301f150905090810190601f1680156113ae5780820380516001836020036101000a031916815260200191505b5091505060006040518083038185876185025a03f19250505015156113d257610002565b60048701805461ff00191661010017905542600255600854875460019950600160a060020a03908116911614156114155786880154600354101561144b57610002565b60018701805430600160a060020a031660009081526006602052604090208054919091019055546007805490910190555b6114ae565b600187015460038054919091039055611446565b811015801561146e5750858510155b156114ae57604051600b8801546006890154600160a060020a0391909116916000919082818181858883f1935050505015156114a957610002565b426002555b60048701805460ff19169055604080518981526020810183905281518c927fdfc78bdca8e3e0b18c16c5c99323c6cb9eb5e00afde190b4e7273f5158702b07928290030190a25050505050505092915050565b61150a3361066f565b6000141561151757610002565b60008054879081101561000257508052600c86027f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e566810154600080516020611fca83398151915290910193504210806115785750600383015462360d800142115b8061159157508254600160a060020a0390811690861614155b806115a15750600783015460ff16155b156115ab57610002565b600883018054600090811015610002579081526020812060030154600160a060020a03161415611712576117fc85604051600090600160a060020a038316907feaa602bff84cc95bf1b06084485d789e35140c4931eeca8a3b6eb0ed87d74917908390a260408051600b547f0b6d8d52000000000000000000000000000000000000000000000000000000008252600160a060020a0380861660048401526024830185905262375f00420160448401529251921691630b6d8d5291606480820192602092909190829003018187876161da5a03f1156100025750506040515191506106899050565b600883018054600685015430600160a060020a0316310391906000908110156100025781815260208082209390935530600160a060020a03168152600690925260408220548154909290811015610002579081526020812090506002015560115460088401805460009081101561000257908152602081209050600101555b6008830180546000908110156100025781548282526020822060010154929190811015610002579081526020812090505433600160a060020a03166000908152600f602052604081205460088701805493909102939093049450908110156100025790815260208120905060030154604080517f13d4bc2400000000000000000000000000000000000000000000000000000000815233600160a060020a039081166004830152915192909116916313d4bc249185916024828101926020929190829003018185886185025a03f11561000257505060405151600014159150611872905057610002565b60088401805460009081101561000257818152602081206003018054600160a060020a03191690931790925580549091908110156100025790815260208120905060030154600160a060020a03166000141561185757610002565b600683015430600160a060020a031631101561169357610002565b6008830180546000908110156100025781548282526020822060010154929190811015610002579081526020812090506002015433600160a060020a03166000908152600f60205260408120546008870180549390910293909304935083926006928290811015610002575260208120815060030154600160a060020a0390811682526020808301939093526040918201600090812080549095019094553016835260069091529020548190101561192957610002565b30600160a060020a0390811660009081526006602090815260408083208054869003905533909316808352600f82528383205484519081529351929390927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef929181900390910190a3600f600050600033600160a060020a031681526020019081526020016000206000505460116000828282505403925050819055506000600f600050600033600160a060020a03168152602001908152602001600020600050819055506001935050505092915050565b33600160a060020a03818116600090815260066020526040808220546011543090941683529120549092611a2e9061066f565b02040191506009600050600033600160a060020a0316815260200190815260200160002060005054600760005054600860009054906101000a9004600160a060020a0316600160a060020a031663d2cc718f6040518160e060020a0281526004018090506020604051808303816000876161da5a03f1156100025750505060405180519060200150840204039050600860009054906101000a9004600160a060020a0316600160a060020a0316630221038a33836040518360e060020a0281526004018083600160a060020a03168152602001828152602001925050506020604051808303816000876161da5a03f1156100025750506040515115159050611b3557610002565b33600160a060020a031660009081526009602052604090208054820190555050565b8015611b735750611b73838360006000341115611e4157610002565b15611b805750600161037d565b610002565b8015611ba25750611ba284848460006000341115611ee757610002565b15611b8057506001611bb3565b5060005b9392505050565b30600160a060020a031633600160a060020a0316141580611be75750600a30600160a060020a0316310481115b15611bf157610002565b600a55565b600454600160a060020a039081163390911614611c1257610002565b60058054600181018083558281838015829011611c4257818360005260206000209182019101611c429190610cda565b5050506000928352506020909120018054600160a060020a03191683179055506001610689565b5060006108c8565b600c5442118015611c855750600e5460ff16155b8015611cb6575033600160a060020a03166000818152600f602052604080822054905182818181858883f193505050505b15611d665733600160a060020a03166000818152600f602090815260409182902054825190815291517fbb28353e4598c3b9199101a66e0989549b659a59a54d2c27fbb183f1932c8e6d9281900390910190a2600f600050600033600160a060020a031681526020019081526020016000206000505460116000828282505403925050819055506000600f600050600033600160a060020a03168152602001908152602001600020600050819055505b565b600160a060020a038616600090815260096020526040902054808502919091049150811115611d9657610002565b600160a060020a038086166000908152600960205260408082208054859003905591861681522080548201905560019150509392505050565b5060005b600554811015611e3057600580548290811015610002576000919091527f036b6384b5eca791c62761152d0c79bb0604c104a5fb6f4eb0703f3154bb3db00154600160a060020a039081169084161415611e395760019150610484565b60009150610484565b600101611dd3565b33600160a060020a03166000908152600f6020526040902054829010801590611e6a5750600082115b15611edf5733600160a060020a039081166000818152600f6020908152604080832080548890039055938716808352918490208054870190558351868152935191937fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef929081900390910190a350600161037d565b50600061037d565b600160a060020a0384166000908152600f6020526040902054829010801590611f305750601060209081526040600081812033600160a060020a03168252909252902054829010155b8015611f3c5750600082115b15611baf57600160a060020a038084166000818152600f6020908152604080832080548801905584891680845281842080548990039055601083528184203390961684529482529182902080548790039055815186815291519293927fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef9281900390910190a3506001611bb356290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e563',
-            gas: 3000000,
-            gasPrice: 500000000000
-   }, function(e, contract){
-    // funny thing, without this geth hangs
-    console.log("At DAO creation callback");
-    if (typeof contract.address != 'undefined') {
-         console.log('DAO Contract mined! address: ' + contract.address + ' transactionHash: ' + contract.transactionHash);
-    }
- });
-        checkWork();
-       }
-   });
-checkWork();
-
-console.log("mining contract, please wait");
