+++ tests/test.py
@@ -9,8 +9,9 @@ import json
 import subprocess
 import shutil
 import sys
-from datetime import datetime
 import calendar
+from datetime import datetime
+from string import Template
 
 
 def is_exe(fpath):
@@ -103,84 +104,20 @@ class TestContext():
 
     def create_deploy_js(self, dabi, dbin, cabi, cbin):
         print("Rewritting deploy.js using the compiled contract...")
-        f = open("deploy.js", "w")
-        f.write("""// geth --networkid 123 --nodiscover --maxpeers 0 --genesis ./genesis_block.json --datadir ./data  js deploy.js 2>>out.log.geth
-
-console.log("unlocking account");
-personal.unlockAccount(
-    web3.eth.accounts[0],
-    "Write here a good, randomly generated, passphrase!"
-);
-
-function checkWork() {
-    if (!eth.mining) {
-       setTimeout(function() {
-            miner.stop(0);
-        }, 3000);
-        miner.start(1);
-    } else {
-        miner.stop(0);
-    }
-    if (eth.getBlock("pending").transactions.length > 0) {
-        if (eth.mining) return;
-        console.log("== Pending transactions! Mining...");
-        miner.start(1);
-    } else {
-        miner.stop(0);  // This param means nothing
-        console.log("== No transactions! Mining stopped.");
-    }
-}
-
-var _defaultServiceProvider = web3.eth.accounts[0];
-
-var daoContract = web3.eth.contract(""")
-        f.write(dabi)
-        f.write(""");
-
-console.log("Creating DAOCreator Contract");
-var creatorContract = web3.eth.contract(""")
-        f.write(cabi)
-        f.write(""");
-var _daoCreatorContract = creatorContract.new({from: web3.eth.accounts[0], data: '""");
-        f.write(cbin)
-        f.write("""',
-    gas: 3000000
-       }, function(e, contract){
-           if (e) {
-               console.log(e+" at DAOCreator creation!");
-           }
-           if (typeof contract.address != 'undefined') {
-               console.log('DAOCreator mined! address: ' + contract.address + ' transactionHash: ' + contract.transactionHash);
-               checkWork();
-               var dao = daoContract.new(
-                   _defaultServiceProvider,
-                   contract.address,""")
-        f.write("""
-                   {},
-                   {}""".format(str(self.min_value), str(self.closing_time)))
-        f.write(""",
-            {
-                from: web3.eth.accounts[0],
-                data: '""")
-        f.write(dbin)
-        f.write(
-            """',
-            gas: 3000000,
-            gasPrice: 500000000000
-   }, function(e, contract){
-    // funny thing, without this geth hangs
-    console.log("At DAO creation callback");
-    if (typeof contract.address != 'undefined') {
-         console.log('DAO Contract mined! address: ' + contract.address + ' transactionHash: ' + contract.transactionHash);
-    }
- });
-        checkWork();
-       }
-   });
-checkWork();
-
-console.log("mining contract, please wait");"""
+        with open(os.path.join(self.tests_dir, 'templates', 'deploy.template.js'), 'r') as f:
+            data = f.read()
+            # data = f.read().replace('\n', '')
+        tmpl = Template(data)
+        s = tmpl.substitute(
+            dao_abi=dabi,
+            dao_bin=dbin,
+            creator_abi=cabi,
+            creator_bin=cbin,
+            min_value=self.min_value,
+            closing_time=self.closing_time
         )
+        with open("deploy.js", "w") as f:
+            f.write(s)
 
     def run_test_deploy(self):
         print("Running the Deploy Test Scenario")
